# Use flattened_for_each in the fill tessellator.