This improves the performance of generating the fill events a bit.
The PR also adds a few flattening benchmark illustrating the difference between the iterator and builder approaches.

Before:

```
test fill_logo_events_and_tess         ... bench:  13,466,828 ns/iter (+/- 1,275,885)
test fill_logo_events_only             ... bench:   5,216,015 ns/iter (+/- 613,182)
```

After:

```
test fill_logo_events_and_tess         ... bench:  12,624,150 ns/iter (+/- 803,110)
test fill_logo_events_only             ... bench:   4,368,951 ns/iter (+/- 1,420,001)
```